Use normalizer classes
The search classes were getting a bit complicated because they were
concerned with both searching and normalizing. Switching to normalizer
classes allows each class to focus on a single concern.
