+++ app/models/search_google.rb
@@ -26,43 +26,10 @@ class SearchGoogle
   # @param term [string] The string we are searching for
   # @return [Hash] A Hash with search metadata and an Array of {Result}s
   def search(term)
-    raw_results = @service.list_cses(
+    @service.list_cses(
       term,
       cx: ENV['GOOGLE_CUSTOM_SEARCH_ID'],
       num: ENV['RESULTS_PER_BOX'] || 3
     )
-    to_result(raw_results)
-  end
-
-  private
-
-  # Translate Google results into local result model
-  def to_result(results)
-    norm = {}
-    norm['total'] = results.queries['request'][0]
-                           .total_results.to_i
-    norm['results'] = []
-    extract_results(results, norm)
-    norm
-  end
-
-  # Extract the information we care about from the raw results and add them
-  # return them as an array of {result}s
-  def extract_results(results, norm)
-    return unless results.items
-    results.items.each do |item|
-      result = Result.new(item.title, year_from_dc_modified(item),
-                          item.link, 'website')
-      norm['results'] << result
-    end
-  end
-
-  # Extract year from dc.modified meta headers if available
-  def year_from_dc_modified(item)
-    return unless item.pagemap
-    return unless item.pagemap['metatags']
-    dc_mod = item.pagemap['metatags'][0]['dc.date.modified']
-    return unless dc_mod
-    Date.parse(dc_mod).year.to_s
   end
 end
